requirements.txt
accelerate==0.31.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.29.2
antlr4-python3-runtime==4.9.3
anyio==4.8.0
appdirs==1.4.4
appnope==0.1.4
astrapy==1.5.2
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
blis==1.2.0
boto3==1.38.2
botocore==1.38.2
bs4==0.0.2
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cryptography==44.0.2
cssselect==1.3.0
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
datasets==3.5.0
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
deprecation==2.1.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
effdet==0.4.1
emoji==2.14.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
eval_type_backport==0.2.2
executing==2.2.0
faiss-cpu==1.10.0
fake-useragent==2.1.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-vision==3.10.1
googleapis-common-protos==1.69.2
graph-rag-example-helpers==0.6.1
graph-retriever==0.6.1
groq==0.20.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
immutabledict==4.2.1
importlib_metadata==8.6.1
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jq==1.8.0
json_repair==0.39.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langchain==0.3.21
langchain-anthropic==0.3.10
langchain-astradb==0.5.3
langchain-community==0.3.20
langchain-core==0.3.45
langchain-experimental==0.3.4
langchain-graph-retriever==0.6.1
langchain-groq==0.3.0
langchain-huggingface==0.1.2
langchain-neo4j==0.4.0
langchain-ollama==0.2.3
langchain-openai==0.3.8
langchain-text-splitters==0.3.7
langchain-unstructured==0.1.6
langcodes==3.5.0
langdetect==1.0.9
langgraph==0.3.21
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.59
langsmith==0.3.13
language_data==1.3.0
layoutparser==0.3.4
libmagic==1.0
lightning-utilities==0.14.3
litellm==1.67.2
loguru==0.7.3
lxml==5.3.1
lxml_html_clean==0.4.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
meta_ai_api==1.2.5
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy-extensions==1.0.0
narwhals==1.30.0
neo4j==5.28.1
neo4j-graphrag==1.6.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
olefile==0.47
ollama==0.4.7
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
openai==1.76.0
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
ormsgpack==1.9.0
packaging==24.2
pandas==2.2.3
pandoc==2.4
parse==1.20.2
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pi_heif==0.22.0
pikepdf==9.5.2
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
plumbum==1.9.0
ply==3.11
poppler-utils==0.1.0
portalocker==2.10.1
preshed==3.0.9
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pyee==11.1.1
Pygments==2.19.1
pymongo==4.11.2
pypandoc==1.15
pyparsing==3.2.1
pypdf==5.4.0
pypdfium2==4.30.1
pyppeteer==2.0.0
pyquery==2.0.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytorch-lightning==2.5.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
qdrant-client==1.14.1
ragas==0.2.14
ragchecker==0.1.9
rank-bm25==0.2.2
RapidFuzz==3.12.2
refchecker==0.2.14
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-html==0.10.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
s3transfer==0.12.0
sacrebleu==2.5.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==76.0.0
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.39
srsly==2.5.1
stack-data==0.6.3
streamlit==1.44.1
streamlit-option-menu==0.4.0
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
thinc==8.3.4
thinc_apple_ops==1.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
torchaudio==2.6.0
torchmetrics==1.7.1
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typer==0.15.2
types-PyYAML==6.0.12.20250326
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
unstructured==0.17.2
unstructured-client==0.31.1
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.15
urllib3==1.26.20
uuid6==2024.7.10
w3lib==2.3.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websockets==10.4
widgetsnbextension==4.0.13
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.2.2
xxhash==3.5.0
yarl==1.18.3
yfiles_jupyter_graphs==1.10.1
zipp==3.21.0
zstandard==0.23.0
